# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/projects": {
      "post": {
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "ProjectController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all projects for a user",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        },
        "summary": "Get a specific project by ID",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        },
        "summary": "Update a project",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "summary": "Delete a project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/work-schedule/{id}": {
      "get": {
        "operationId": "WorkScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSchedule"
                }
              }
            }
          }
        },
        "summary": "Buscar agenda por ID",
        "tags": [
          "Work Schedule"
        ]
      },
      "patch": {
        "operationId": "WorkScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agenda atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkSchedule"
                }
              }
            }
          }
        },
        "summary": "Atualizar agenda",
        "tags": [
          "Work Schedule"
        ]
      },
      "delete": {
        "operationId": "WorkScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da agenda",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agenda removida"
          }
        },
        "summary": "Remover agenda",
        "tags": [
          "Work Schedule"
        ]
      }
    },
...
    "/forecast/generate": {
      "post": {
        "operationId": "ForecastController_generateForecast",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateForecastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Previsão gerada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            }
          }
        },
        "summary": "Gerar previsão mensal de trabalho",
        "tags": [
          "Forecast"
        ]
      }
    },
...
    "/forecast/{id}": {
      "get": {
        "operationId": "ForecastController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da previsão",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Previsão encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            }
          }
        },
        "summary": "Buscar previsão por ID",
        "tags": [
          "Forecast"
        ]
      }
    },
    "/forecast/work-schedule/{workScheduleId}": {
      "get": {
        "operationId": "ForecastController_findByWorkSchedule",
        "parameters": [
          {
            "name": "workScheduleId",
            "required": true,
            "in": "path",
            "description": "ID da agenda de trabalho",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Previsões da agenda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Forecast"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar previsões por agenda de trabalho",
        "tags": [
          "Forecast"
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/github": {
      "get": {
        "operationId": "AuthController_getGitHubAuthUrl",
        "parameters": [
          {
            "name": "redirect_uri",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Work Schedule API",
    "description": "API para gerenciamento de agendas de trabalho",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "E-commerce Website"
          },
          "description": {
            "type": "string",
            "description": "Project description",
            "example": "Building a modern e-commerce platform"
          },
          "hourlyRate": {
            "type": "number",
            "description": "Hourly rate for this project",
            "example": 50
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "default": "USD",
            "example": "USD"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "COMPLETED",
              "PAUSED",
              "CANCELLED"
            ],
            "description": "Project status",
            "default": "ACTIVE"
          },
          "startDate": {
            "type": "string",
            "description": "Project start date",
            "example": "2024-01-15"
          },
          "endDate": {
            "type": "string",
            "description": "Project end date",
            "example": "2024-06-15"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          }
        },
        "required": [
          "name",
          "hourlyRate",
          "currency",
          "status",
          "userId"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {}
      },
...
      "WorkSchedule": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkScheduleDto": {
        "type": "object",
        "properties": {}
      },
...
      "GenerateForecastDto": {
        "type": "object",
        "properties": {}
      },
      "Forecast": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}